+++ src/ipc/modules/persistence/importObject/importObject.ts
@@ -1,6 +1,6 @@
 import fs from 'fs';
 import path from 'path';
-import getAppDataDirectory from '_/ipc/utils/persistence/getAppDataDirectory';
+import { directories } from '../../../utils/persistence/directories';
 
 /**
  * Imports the specified json file and parses it to an object of type T
@@ -8,7 +8,7 @@ import getAppDataDirectory from '_/ipc/utils/persistence/getAppDataDirectory';
  * @returns `null` if the import failed, otherwise the imported object
  */
 export function importObject<T>(filename: string): T | null {
-  const inputPath = path.join(getAppDataDirectory(), filename);
+  const inputPath = path.join(directories.appData(), filename);
   if (!fs.existsSync(inputPath)) {
     return null;
   }