+++ src/ipc/modules/persistence/uploadDocument/uploadDocument.spec.ts
@@ -1,8 +1,8 @@
 import fs from 'fs';
 import path from 'path';
 import { v4 } from 'uuid';
+import { directories } from '../../../utils/persistence/directories';
 import { DocumentTarget } from '../../../utils/persistence/documentTarget';
-import * as getAppDataDirectoryModule from '../../../utils/persistence/getAppDataDirectory';
 import uploadDocument from './uploadDocument';
 
 jest.mock('uuid', () => ({
@@ -10,7 +10,6 @@ jest.mock('uuid', () => ({
 }));
 
 describe('uploadDocument', () => {
-  const appDataDir = 'appData';
   const uploadedFile = 'test/file.txt';
   const target: DocumentTarget = {
     type: 'resident',
@@ -18,18 +17,20 @@ describe('uploadDocument', () => {
   };
   const expectedDocumentId = 'uploaded-document-id';
 
-  const destinationDir = path.join(appDataDir, `documents/residents/${target.residentId}`);
-  const destinationFile = path.join(destinationDir, `${expectedDocumentId}.txt`);
+  const destinationDir = path.join(
+    directories.appData(),
+    `documents/residents/${target.residentId}`,
+  );
+  const destinationFile = path.join(
+    destinationDir,
+    `${expectedDocumentId}.txt`,
+  );
 
   let existsSyncMock: jest.SpyInstance;
   let mkdirSyncMock: jest.SpyInstance;
   let copyFileSyncMock: jest.SpyInstance;
 
   beforeEach(() => {
-    jest
-      .spyOn(getAppDataDirectoryModule, 'default')
-      .mockReturnValue(appDataDir);
-
     existsSyncMock = jest.spyOn(fs, 'existsSync').mockReturnValue(true);
     mkdirSyncMock = jest.spyOn(fs, 'mkdirSync').mockImplementation();
     copyFileSyncMock = jest.spyOn(fs, 'copyFileSync').mockImplementation();