+++ src/ipc/utils/persistence/uploadMarkdownAsPdf/uploadMarkdownAsPdf.ts
@@ -2,8 +2,8 @@ import { rm } from 'fs/promises';
 import path from 'path';
 import { mdToPdfFile } from 'electron-md-to-pdf';
 import { v4 as uuid } from 'uuid';
+import { directories } from '../directories';
 import { DocumentTarget } from '../documentTarget';
-import { getTmpDirectory } from '../getAppDataDirectory';
 import uploadDocument from '_/ipc/modules/persistence/uploadDocument/uploadDocument';
 
 type UploadMarkdownAsPdfArgs = {
@@ -21,7 +21,7 @@ type UploadMarkdownAsPdfArgs = {
 export async function uploadMarkdownAsPdf(
   args: UploadMarkdownAsPdfArgs,
 ): Promise<string> {
-  const tmpFile = path.join(getTmpDirectory(), `${uuid()}.pdf`);
+  const tmpFile = path.join(directories.temporary(), `${uuid()}.pdf`);
   await mdToPdfFile(args.markdownContent, tmpFile, {
     ...args.mdToPdfOptions,
     showdownOptions: {